+++ test/repo_analyzer/analyze_test.clj
@@ -3,7 +3,7 @@
             [repo-analyzer.analyze :refer :all]
             [clojure.tools.trace :refer :all]))
 
-(deftest test-analyze-repository-integrity-commit-statistics
+(deftest test-analyze-repository-commit-statistics-integrity
   (testing "Checks the commit-statistics analyze-repository function using the project's own repository by
   checking some simple constraints"
     (let [result (analyze-repository ".")]