+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file. This change
 ## [Unreleased]
 ### Added
 - Information about contributors
+- Information about number of commits
 ### Changed
+- Include bootstrap for HTML output
 ### Removed
 ### Fixed