+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file. This change
 - Information about self-committed commits
 - Metainformation about analysis
 - Information about commits with different author/committer
+- Printouts for analysis steps
 ### Changed
 - Include bootstrap for HTML output
 ### Removed